optimize: switch to SSR, enable image optimization, and lazy load components
