+++ package.json
@@ -74,6 +74,7 @@
 	},
 	"devDependencies": {
 		"@eslint/js": "^9.39.2",
+		"@next/bundle-analyzer": "^16.1.2",
 		"@tailwindcss/typography": "^0.5.19",
 		"@types/aos": "^3.0.7",
 		"@types/node": "^22.19.5",
